Updated Expo SDK version from 52 to 53 so the app can run on the latest Expo Go.
Expo have recently release a new version of the Expo SDK (53) that is compatible with the latest Expo Go app. After the upgrade, the app failed to recognise route components correctly and caused errors.
To resolve this, I added a custom `metro.config.js` file to explicitly include the correct file extensions and ensure the Metro bundler resolves files as expected. This restores proper routing behaviour and allows the app to compile and run correctly.